Make sure when syncing to a branch, we pick up new subrepos in it.
Summary:
Desmond was seeing this problem, where he introduced a new subrepo in
his branch, but we weren't picking it up in the deploy script because
we only run 'git submodule sync/init/etc' in the *master* branch on our
fancy repo-for-everyone dir.

Now we change to the right repo before doing the submodule
operations.  I'm not super-happy with this because (a) 'git checkout'
can fail when repos aren't clean, which I feel can happen to us, and
(b) we do a bunch of 'git clean's which don't make much sense now
since the list of submodules may differ from branch to branch.  I
//think// this means we'll just be doing unnecessary clean's as we go
back and forth between repos, but it's hard for me to get my head
around.

In any case, this is better than the status quo.  Hopefully.

Test Plan: Fingers crossed.

Reviewers: chris

Reviewed By: chris

Subscribers: desmond

Differential Revision: https://phabricator.khanacademy.org/D15846