Don't try to use 'git new-workdir' for most subrepos.
It's not really safe to call git new-workdir on each submodule, since
it doesn't deal well with submodules appearing and disappearing
between branches.  (I had some attempt to deal with this by changing
branches in the repos-dir before syncing subrepos, but it was a big
mess and doesn't seem to be working 100% correctly.)

So now we hard-code a few of the big submodules that have been around
a long time and aren't going anywhere, and use git new-workdir on
those, and use 'normal' submodules for everything else.

Auditors: chris

Test plan:
Ran the following set of commands on my local machine.

    cd /tmp
    rm -rf webapp2
    git new-workdir $HOME/khan/webapp webapp2
    export REPOS_ROOT=$HOME/khan/webapp
    source ~/khan/jenkins-tools/build.lib
    cd webapp2
    _safe_update_submodules intl/translations third_party/polib-khansrc
    _safe_update_submodules intl/translations third_party/polib-khansrc
    _safe_update_submodules khan-exercises
    _safe_update_submodules third_party/javascript-khansrc/rainbow
    _safe_update_submodules third_party
    _safe_update_submodules third_party
    _safe_update_submodules

It updated submodules correctly in each case.