Don't error, if old rest and datastore files don't exist & fix en-pt call.
Summary:
This handles the changes that are needed for D16491.  New languages
don't have rest or datastore files, so we don't want to die if we fail
to remove them.

en-pt will now b lower cased in keeping with ka_locale convention.

Test Plan:
cross-fingers (I ran a copied out and altered subset of
update-translation locally with the right DATA_DIR and it seemed to
work).

Reviewers: csilvers

Subscribers: kevinb, jlfwong

Differential Revision: https://phabricator.khanacademy.org/D16493