Modify build.lib to support 'git bigfile', and use it in update-translations.
Summary:
The change to update-translations is additive: we now push all
translation updates to two repos: the existing repo, and a new repo
that is set up to use git-bigfile (because I checked in a
.gitattributes file in it).

build.lib now does 'git bigfile push' and 'git bigfile pull' in the
places where it's needed.

This was easier to implement by re-combining
checkin-update-translations and update-translations.  They were
originally separated for use by a control-flow we don't need anymore,
so combining them back should be fine.

Test Plan: Fingers way crossed.  We'll see how the next update-translations run goes!

Reviewers: james, chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D17323