Delete old bigfiles every time we fetch new ones.
Summary:
By default, 'git bigfile pull' keeps all the old bigfiles around on
the local filesystem.  This is useful when you 'git checkout' back in
time: the objects are already here.  But we don't do do
backwards-checkouts on jenkins, and the old bigfiles take up a lot of
space.  So we prune them.

I made it so we delete all bigfiles fetched 2 or more days ago.  That
should be plenty of time to use the bigfile before moving on into the
future.  If we *do* need an old bigfile again, no problem we'll just
fetch it again from S3.

Test Plan: Fingers crossed.

Reviewers: james

Reviewed By: james

Subscribers: jlfwong, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D17696