Stop breaking up pofiles into smaller files.
Summary:
We had been breaking up pofiles because github had a 100MB limit.  These
files are in bigfile now, so we don't need to.  We can also remove
breaking things up into rest and datastore, but that is a bigger change
and can be done in a different commit.

Test Plan: cross-fingers

Reviewers: csilvers, kevinb

Differential Revision: https://phabricator.khanacademy.org/D17860

Conflicts:
	update-translations.sh