Add a jenkins script for uploading graphie labels
Summary:
This is a script to be run by a new upload-graphie-labels job
on jenkins. It performs the job of uploading all of the translated
graphie labels back to S3 using the upload_graphie_labels.py script.

I copied most of this out of the update-translations.sh script.

Test Plan: - Not sure how to test this. Should I just try running the script?

Reviewers: james

Reviewed By: james

Differential Revision: https://phabricator.khanacademy.org/D17960