We use word counts for displaying translation analytics at http://www.khanacademy.org/translations/es/analytics We are moving away from calculating the word counts at deploy time (since it takes 6.5 mins on jenkins) and doing it instead in update...
Summary:
...-translations, then checking it into intl/translations subrepo.

I've added a copy of the latest word_counts file here:
https://phabricator.khanacademy.org/rGWIB7e534685a96a
So that the code can always expect the file to exist and don't see any dips in
the data.

I've added the code to cacluate the word_counts.json file and update
where we are trying to read it from here:
https://phabricator.khanacademy.org/D18510

This diff uses the make rule from D18510 for calculating the word counts and
copies the file over to intl/translations.

Test Plan: cross-fingers for the next update-translations run.

Reviewers: kevinb

Subscribers: csilvers, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D18513