When resetting, reset submodules as well.
Summary:
We periodically see errors in the jenkins logs that a submodule is
dirty, causing the script to fail.  See, e.g.
   http://jenkins.khanacademy.org/job/update-translations/1379/console

   timeout 60m git submodule update --init --recursive [...]
   [...]
   error: Your local changes to the following files would be overwritten by checkout:
   pofiles/accents.datastore.po
   pofiles/accents.rest.po
   pofiles/bg.datastore.po
   pofiles/bg.rest.po
   pofiles/bn.datastore.po
   [...]

I believe this is because the _safe_destructive_checkout that preceeds
it is not actually totally safe.  I think doing the 'reset --hard' in
the submodules as well as in the main module *will* make it safe, so
these errors shouldn't recur.

Test Plan: Fingers crossed

Reviewers: chris, james, ethan

Reviewed By: james, ethan

Subscribers: csilvers, james

Differential Revision: https://phabricator.khanacademy.org/D18571