Download approved pofiles first.
Summary:
Unfortuantely as we have to do two different exports of pofiles, we get
two different snapshots of translations efforts at different times.  It
is a bit saner to reason that if a translation is approved, it is also
translated.  This hadn't been the case as it seems some approvers might
also translate and then approve their own translation in the time
between the two different exports.

Test Plan: cross-fingers

Reviewers: csilvers, jlfwong

Subscribers: chelsea, kevinb, csilvers, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D18755