Make safe_sync_to_origin have the local branch match up with origin.
Before, it would just sync us to origin/branchname, which was fine as
far as it went, but sometimes other code would assume we were checked
out on a local branch, and do things like 'git rev-parse <branch>'.
It was not unreasonable for this to work, since they had called
'safe_sync_to_origin <branch>' prior to this.

Now we make it so that *does* work by making 'safe_sync_to_origin' not
only sync to 'origin/branch' but sets up 'branch' to be the same as
'origin/branch'.

Auditors: ethan, jlfwong

Test plan:
Will do a commit after this!