Re-raise any unknown errors in set-default
Summary:
I'm not entirely sure if this is the right thing to do, but I think it is.
Right now, if we get an unhandled error in set-default, we log it, and send a
failure message to hipchat, but Jenkins sees it as a success.  This seems
suboptimal to me, and caused some confusion when we got 409s from Google last
week.  With this change, we'll mark it as a failure, since it means the
set-default job probably didn't complete correctly.  (Monitoring errors will
continue to mark the build a success, since those are only a failure if a human
deems it so.)

Test Plan: Fingers crossed.

Reviewers: jlfwong, chris

Reviewed By: chris

Subscribers: alan, csilvers

Differential Revision: https://phabricator.khanacademy.org/D20587