Revert "Re-raise any unknown errors in set-default"
Summary:
After talking to @csilvers, it turns out that we don't want the build marked as
a failure unless we will also abort and roll back.  We may not want to do that,
so we'll leave the build a success and Mr Gorilla's message telling people it
failed and what the next steps are.

This reverts commit ac78a8da489438b6e8824ad3bf928301114a5897.

Reviewers: chris, csilvers

Reviewed By: chris, csilvers

Subscribers: csilvers, alan, jlfwong

Differential Revision: https://phabricator.khanacademy.org/D20953