Remove unused webapp-i18n-bigfile repo copy
Summary:
This appears to be legacy from when we had the bigfile and non-bigfile
versions of repos around. Bigfile won and we no longer use the other
repo. The evidence is that in update-translations job logs on Jenkins,
the second commit did a bunch of merging and exited early with
"Everything up-to-date".

Once the current update-translations job finished, I'll go into Jenkins
and manually delete the webapp-i18n-bigfile clone in the workspace.

Test Plan: Fingers crossed.

Reviewers: james, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D20961