modularize manual_test
Summary:
This breaks down generation of plaintext, HTML (when needed), and Slack
attachment into variables, which are then sent independently.

This adds a plaintext output for this action (previously HTML was sent
to logs which would be message), which is also used as fallback for the
Slack attachments.

I wouldn’t suggest doing this separation with all actions, but for the
complex ones where we have specific logic I believe it makes it easier.

For discussion!

Test Plan:
I’m a little nervous about this so @bmp if you could please help me
test locally I’m not sure how it works in your setup.

Reviewers: benkraft, benjaminpollack

Reviewed By: benjaminpollack

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21073