Lightweight split of update-translations.sh
Summary:
This is a cheap 'n easy split so that we can get closer to playing with
the Jenkins job settings to run the job in various modes (download only,
update only, or some combo).

When this lands, I'll update the job config on Jenkins to pass in both
flags so the job should be functionally the same as it is now.

I didn't indent the code within the new if statements to avoid a
hard-to-read diff. I can do an indent-only commit in the future.

Test Plan:
Ran `bash update-translations.sh` and got the error about setting one of
the flags.

Reviewers: james

Reviewed By: james

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21028