Write jenkins props to a JSON for Sun
Summary:
I put it in the current lockdir.  I think this is the right thing to do; if
we're hitting a custom lockdir, Sun's going to get confused anyway, as will he
if something funky with relocking is happening.  So for sun's purposes, as long
as there's no lockdir in the standard location, we don't need to worry that a
deploy is going on.

Test Plan: Fingers crossed.

Reviewers: benjaminpollack, csilvers, mroth

Reviewed By: mroth

Differential Revision: https://phabricator.khanacademy.org/D21099