Log whether the job is actively running
Summary:
This is an attempt to better support Hubot using Jenkins as the state
master. While this patch is not perfect, it also is, IMVHO, //no worse//
than what Sun is currently doing. We can have the script marked running
by accident, but not have it not marked running, and clearing it is
easy.

Test Plan: YOLO

Reviewers: mroth, csilvers, benkraft

Reviewed By: mroth, csilvers, benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21214