nicer format for deploy start messages
Summary:
This is the most common message in the deploy process, so wanted to
clean it up a bit.

I tested this message in mockup but I had to do a lot of munging to get
it to work with our existing code path, so I’m almost certain I did
something dumb here.  Sanity check requested.

Also, the tree spinner is not retina capable yet, so if we land this we
are pretty committed to someone going into photoshop for a bit.

Test Plan: Fingers crossed?

Reviewers: benjaminpollack, benkraft

Reviewed By: benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21390