Send test failures to slack.
Summary:
Since this is a part of the deploy process, make it a little pretty.  I changed
the sender because it would be slightly annoying to make the messages match
Jenkins's (as they currently do) and that doesn't make sense anyway, since
they're not from Jenkins itself.

Test Plan: fingers very crossed.

Reviewers: mroth, csilvers

Subscribers: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D21474