Add option to control language download in update-translations.sh
It'd be nice if the Jenkins job could be invoked manually with
particular languages to process, for testing or because in some case,
the humans know better than the prioritizer.

Auditors: james

Test Plan:
I ran the new lines within an invocation of /bin/sh and they worked as
expected. So, somewhat fingers crossed.