Add link to logs in the "jenkins noticed a problem" message
Summary:
Right now it's just a link to the logs for the default module.  Once we kill
hipchat, it will be easier to link to the right module, where there is an error
from a specific module; right now that would require a bunch of plumbing since
we take that text verbatim from set_default.py and dump it into both slack and
hipchat.

Test Plan: fingers crossed.

Reviewers: benjaminpollack, alan, mroth

Reviewed By: mroth

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21803