Remove HipChat from deploy.sh
Summary:
Setting `--slack-channel` had never actually been added, so I did that too.
(It was previously the default, which we ~never override.)

Since both args have a default, this can be deployed independently of the
webapp changes, although it will likely make the most sense to deploy it
afterwards.

Test Plan: fingers crossed.

Reviewers: benjaminpollack, mroth

Reviewed By: mroth

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D21881