Add E2E test analysis to analyze_make_output
Summary:
This commit adds an option to run analyze_make_output on the output of an
end-to-end test run.  This will be used for running end-to-end tests on Jenkins
on every deploy.

This also makes a change to the API of this script: all files it should process
must be specified explicitly.  This is because E2E tests will be run separately
from the rest of the tests.  All callers on Jenkins already specify the input
files explicitly, so this does not require any updates to them.

Test Plan: Run it locally on local E2E output.

Reviewers: sean, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D22768