Uniquify errors before reporting them.
Summary:
The way the jenkins test-runner works, when a new test is added to our
test-suite, it gets run by all 4 of the worker machines.  If it fails,
it will fail 4 times, and 4 failure messages were getting reported.

This led to spammy failure output (see
https://khanacademy.slack.com/archives/1s-and-0s-deploys/p1446229216007689
), where we listed the same error message 4 times.  Using a set should
avoid this.

Test Plan:
Fingers crossed -- we don't have great tests for jenkins-tools.  Next
time we have a deploy failure due to tests, we'll at least be able to
verify this didn't break anything.

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D23354