Error to chat if we are passed an invalid revision
We check lower down that unless it's ahead of main, `$git_revision` is a valid
branch, but if it's not even a valid git revision, the `checkout` will fail,
causing the deploy to fail silently.  Ideally we'd have a better mechanism of
dealing with a failure before the deploy really starts, but we don't, so for
now we'll just make sure we report properly in this common error case.

Test plan:
Push this, then:
* Start a deploy with an invalid branch and make sure it reports the failure.
* Start a deploy with a valid branch and make sure it merges from master and
  starts the deploy.

Auditors: csilvers