Modify all the i18n jenkins jobs to do their work in the translations branch.
Summary:
The deploy script is already set up to merge the translations branch
when deploying.

This way we move from a model where the i18n jobs can update the
master branch, to one where only the deploy script updates the master
branch.  The i18n jobs all work in a separate branch, that the deploy
process automatically knows to merge in.

The advantages of this scheme are:
1) It allows us to "turn off" merging in translations for a deploy
   where we need it to go a bit faster.
2) It makes the master branch's history a bit nicer by only updating
   intl/translations when we actually deploy, not over and over again.

NOTE: after updating this, I'll also need to change
   https://jenkins.khanacademy.org/job/i18n-gcs-upload/configure
so the default GIT_REVISION is `translations`.  This is the only i18n
job I found that's not controlled by a script in this repo.

Fix a logic bug where we were calculating the branchname in the wrong place.

I still don't like that we need to calculate this branchname at all. :-(

Test Plan:
Fingers very crossed -- this stuff doesn't have good testing.  I'll
have to carefully test jenkins jobs for a while after this to make
sure they're still working right.

(none new)

Reviewers: james, chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D26080