Clean up genfiles every week.
This reduced the size of the deploy-znd repo by 16G!

This rule used to just live in the upload-gcs-files config, but I
figured it made sense to do it everywhere.  (Now it's a bit worse for
upload-gcs-files, since before it would get cleaned every run and now
it's just cleaned once a week.  But should still be good enough.)

Auditors: chris

Test plan:
Ran `clean_genfiles` manually on jenkins as the jenkins user and it
succeeded with exit code 0.