Extract the update-submodule-pointer logic into its own function.
deploy-via-multijob needs to update the intl/translations submodule
pointer, and was using safe_commit_and_push to do it.  That was both
confusing and fragile: confusing because there was never anything to
commit, and fragile because it does a push of the intl/translations
submodule, which can fail if intl/translations was updated (by someone
else) at the wrong time.

deploy-via-multijob doesn't need the commit *or* the push.  So I
extract out the functionality it does need, so we can use it.

Auditors: benkraft

Test plan:
Fingers crossed