Created android_db_generator.sh
Summary:
This will be used from Jenkins, as part of our e2e tests, to be able to verify for each deploy that the updated content or webapp code does not break our mobile clients.
See D27004 for more details on the Android script.

Depends on D27004.

Test Plan: I still haven't been able to test this. We'll have to test in production...

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: #android, benkomalo, #ios

Differential Revision: https://phabricator.khanacademy.org/D27007