Get rid of support for `all`.
James suggested I do this last time, since it's a) unused, and b) too
memory intensive to ever actually be used, but I was like, "what the
heck, let's keep it."  Now I want to update this script to do more
stuff, and having to support `all` is a pain.  So let's get rid of it
after all.

Auditors: james

Test plan:
Fingers crossed