Do not bother to do the 'git clean' and 'git reset' if workspace is clean.
Summary:
This may save a bit of time on every jenkins job.  (Or it may not,
since the main cost of the clean/reset is probably traversing the
filesystem, and the new `git status` we put in will have to do the
same thing.)

Test Plan:
Will check some jenkins jobs after this runs to make sure all looks
good.

Reviewers: chris

Reviewed By: chris

Subscribers: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D28139