Add `make deps` to some scripts that lack it.
Summary:
If you run something that lives in webapp, you never know when it
might start depending on numpy or the like.  We should make sure all
the deps are up to date before running things in weabpp.

I did not change e2e-tests.sh since that mostly runs javascript code,
with very small amount of python, and `make deps` is only for python
deps.  It's possible `make npm_deps` is called for here, but it seems
to be working fine already and I don't want to slow down an already
critical-path operation.

Test Plan: Fingers crossed.

Reviewers: benkraft, chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D28709