Get rid of the `prime` call now that deploy.py does it.
Summary:
I think this call was added just so we'd do content-priming for the
first e2e run during the deploy process (when we're running on the
non-default).  But now we do non-default content-priming as part of
deploy.py, so there's no need to do it here as well.  It just slows
things down (a bit).

Test Plan: Fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D28710