Don't alert *every* time we see some dead workers.
Summary:
When all the test machines are down, and someone starts a new deploy,
the test machines all start up again to run the deploy-tests.  But
they don't all start up at exactly the same time, and it's pretty easy
for the check-live-workers script to run during this startup time, and
see some workers alive and some dead, and complain.

So now I require that we see this bad state for at least 5 minutes
before complaining.  This should be enough time for all the
test-machines to start up.

Test Plan: Fingers crossed.

Reviewers: alexanderforsyth, dhruv

Reviewed By: dhruv

Differential Revision: https://phabricator.khanacademy.org/D28833