Add a script to ensure we send to slack on jenkins errors.
Summary:
This will run daily, so we can discover misconfigured jenkins jobs.

With a few exceptions, we require all jenkins jobs to notify a slack
room on failure.  Otherwise, we have no way of knowing when our jobs,
which we think are working, actuallly are not.

Test Plan:
I ran the script manually except for the sending to slack, and saw
that `$errors` had stuff like this:
```
deploy-react-render-server: should send to slack on 'NotBuilt'
deploy-react-render-server: should send to slack on 'Unstable'
e2e-crawler-test: should send to slack on 'NotBuilt'
e2e-test: should send to slack on 'NotBuilt'
```
I fixed all those errors in jenkins and re-ran the script and saw that
`$errors` is now empty.

Reviewers: dhruv, alexanderforsyth

Differential Revision: https://phabricator.khanacademy.org/D28854