Don't warn until a machine has been down for at least 10 minutes.
The workers get a bit out of sync depending on how long it takes them
to finish running their particular batch of tests.  This means they go
offline at slightly different times.  Usually the difference is just a
few minutes, but it's not uncommon for it to be more than 5 minutes.
So up the monitoring threshold to 10.

Auditors: dhruv

Test plan:
None