Move jenkins worker alerts to #monitoring for now
Summary:
There seems to be some issue that is causing this alert to fire more
often than it should. It's hard to diagnose why, since the alerts seem
to fire in the very early morning and state is lost by the time we get
into the office.

This patch moves the alerts to #monitoring so we don't spam everyone,
but also get notified so we can fix it long term.

Test Plan: Fingers crossed

Auditors: benkraft, csilvers