Use the pickle file and not the debug file.
Summary:
Calling it a debug file is a bit misleading, and led to the unfortunate
series of events which caused: https://app.asana.com/0/27216215224639/187031714959555

I'm not sure why we were using the text file, but it could be that
jenkins was OOM with it and not the text file if I am remembering
correctly.  But we have also increased jenkins memory since then.  At
any rate probably worth the try

Test Plan: cross-fingers

Reviewers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D31071