Get rid of the `selenium_tests` arg to `selenium_webapp_testing.py`.
We want to be running *all* tests here, which means we should rely on
the default behavior, rather than specifying a directory explicitly.
That way, if we want to add tests to a new location, when we update
selenium_webapp_testing.py it will automatically get picked up here.

(It also means things won't break if we rename the selenium_testing
dir :-) )

Auditors: dhruv, johnsullivan

Test plan:
I manually ran:
   tools/selenium_webapp_testing.py --buffer-output --url https://www.khanacademy.org --driver sauce --jobs 4 --retries 3 --xml-dir /tmp/selenium_test_reports
(which is equivalent to what e2e-tests.sh runs, but with the
`selenium_tests` arg gone), and it ran all the tests still.