Separate out the e2e tests into separate scripts.
Summary:
Now that I am planning to run the selenium tests on the jenkins worker
machines, we need to separate out the running of the selenium tests
with the running of the android test, which can run on the main
jenkins, and doesn't fit well into the existing e2e test framework.

Test Plan:
Fingers mostly crossed.  Hard to test locally since `env` doesn't
exist, etc.

Reviewers: dhruv, amos, aasmund

Reviewed By: dhruv

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D33125