Add the ability to run the functions in here from the commandline.
This is useful for groovy scripts, which want to do something like:
   sh './build.lib safe_sync_to <repo> master'

(The alternative is to do something like:
   sh 'sh -c ". ./build.lib; safe_sync_to <repo> master>"'
; you can see why having an executable version is cleaner.)

Note that when run as an executable, the restriction still applies
that you have to run it from a jenkins workspace directory or the
script will die with an error.

Auditors: amos, colin

Test plan:
Ran
   jenkins-tools/build.lib busy_wait_on_dropbox jenkins-tools
from my local `khan` directory, and it failed like so:
   jenkins-tools/build.lib: 626: jenkins-tools/build.lib: dropbox.py: not found
which is just what running the function after sourcing would do!