Play around with defining new steps.
The 'pip' step runs the given shell command in a virtualenv.  I'm
guessing that's going to be the most popular one for us.  I also
always put timestamps in that case, because we always want timestamps
everywhere I'm guessing.

Auditors: amos, chris, colin

Test plan:
Fingers very crossed.  I'm not hopeful on the first try on this one.