Another big code refactor: put a lot of logic into our new `node` wrappers.
I realized setting virtualenv in Setup isn't the right thing to do,
since virtualenv is inherently a workspace-specific action, so it
needs to happen in code that deals with workspaces.  That means inside
each call to `node`.  Lots of other stuff happens inside each call to
`node` as well -- for instance, turning on timestamping -- so it made
sense to write a wrapper for all of it.

I think the resulting code is starting to look pretty clean (though
the helpers have a lot of duplication at the moment).

Auditors: chris, amos, colin

Test Plan:
Totally untested, but
   groovy jobs/e2e-test.groovy
shows no syntax errors.