Fix a bug where I was trying to use a variable out of scope.
I don't quite understand how the scoping rules in groovy relate to the
pipeline DSL, but it (now) seems clear that the `{` and `}` for step
clouse (like `withEnv`) create a new groovy scope.

Auditors: amos, chris, colin

Test Plan:
Ran this change on jenkins and saw it set up the virtualenv properly.