Don't try to combine `body` with optional args.
There is no documentation, and no examples, about how to use define a
new pipeline step that takes both a body and arguments.  Is the body
added as the first arg?  The last?  An arg named `body`?  It's doubly
unclear when the other arguments are optional.

That's exactly the situation I was in, and it exactly wasn't working.
So now I got rid of those arguments, and make people just call a
separate `with` clause if they want secrets.  It causes a bit more
indentation, but on the plus side lets us better control exactly where
secrets can be accessed from.

Auditors: chris, colin, amos

Test Plan:
Fingers crossed, as always.