Use currying to avoid evaluating a variable too late.
Just like in python you need to do the hacky:
    double = lambda i=i: i * 2
to make sure that i is evaluated when double is defined, not when it's
called, you need to do something similar in groovy.

Without this, the i inside our jobs['ec test 0'] closure was being
evaluated when the job started running, which was after i was already
equal to `NUM_WORKER_MACHINES`.  Instead, we wanted it to have the
value 0, since this was our first time through the for loop.  Currying
is the way to do that.

I called the curry-variable `i` as well, as is customary (see the
`i=i` in the python example), though perhaps confusing to someone
trying to figure out what's going on.

Auditors: amos, chris, colin

Test plan:
I feel like we're getting close!