`curry` doesn't work -- I need to use another workaround.
Apparently this is due to a limitation in pipelines?  I'm not sure I
understand the problem or the workaround, but I //think// I'm
following the recommendation in
   https://jenkins.io/doc/pipeline/examples/#parallel-from-list

Auditors: colin, chris, amos

Test plan:
Fingers crossed.  Was playing around with the groovy editor on
jenkins, but it doesn't seem to match what the pipeline DSL does.