Speed up the e2e tests by doing more work in parallel.
While we're syncing webapp on master, we can be syncing on the worker
machines as well.  That saves time since we don't have to sync later!

To maximize the chances our 'pre-sync' will succeed, I added a
throttle-category to keep other jobs from trying to run on the
test-workers while we're running.  Otherwise, it's possible another
job could swoop in between the time we did the pre-sync and the time
we actually started our e2e tests.  We check for that, so it wouldn't
result in an error, but it would be more inefficient.

I changed our existing make-check jobs to use that throttle-category;
they're the only other jobs that use the test-workers currently.

I'm a bit unhappy with isAtCommit, since I wish it could tell that the
workspace was clean, and then we could do that automatically in
build.lib rather than having to do an explicit check.  Maybe in the
future.

Auditors: amos, chris, colin

Test Plan:
Fingers crossed.