Add a mandatory timeout field for all onFoo() calls.
Using the new withTimeout step -- a fancier version of the built-in
`timeout` step, though not as fancy as it will one day be -- we now
specify every time we do any non-trivial work, an upper bound of how
long we expect that work to take.  That will keep this script from
getting caught in an infinite loop and using up executors.

Auditors: amos, chris, colin

Test Plan:
Fingers crossed!  It turned out to be a pretty big bunch of code.  I
extracted parseTimestamp into its own file and ran
   grooy test.groovy
on the commandline a few times, to test parsing various values.