Add a small script to run commands in parallel, and use it in e2e-test.
Summary:
You give the script a command and a bunch of arguments, and it runs
the command multiple times, one with each argument, in parallel.  Its
return value is 0 iff all the parallel processes succeeded.

GNU parallel can do this, but it's a bit overkill and not installed on
unixes by default.  I'm a bit surprised there's no unix tool for this
actually, but there you go.

I may have been better off using xargs syntax (with -iFOO), but this
has simpler semantics and works for our needs.

Test Plan:
I ran the following:
   % time ./in_parallel.py 3 4 5 -- sleep; echo $?
   0:05.01
   0

   % time ./in_parallel.py --failfast 3 4 5 -- sleep; echo $?
   0:05.02
   0

   % time ./in_parallel.py 3 4 5 -- sh -c 'sleep $1' --; echo $?
   0:05.01
   0

   % time ./in_parallel.py 3 4 5 -- sh -c 'sleep $1; exit 1' --; echo $?
   0:05.01
   1

   % time ./in_parallel.py --failfast 3 4 5 -- sh -c 'sleep $1; exit 1' --; echo $?
   0:03.01
   1

which all look right to me.  (We need that extra `--` after the `sh`
the same reason `git alias` does -- sh doesn't count its first
argument as `$1` for some reason, it just drops it on the ground.)

The test for e2e-test.groovy, of course, will be in the running.

Reviewers: amos, colin, chris

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D33482