Get rid of isAtCommit -- it's not safe yet.
The problem is that GIT_REVISION is probably a label (tag or branch),
such as `master`, and we really need to be doing
`isAtCommit(origin/master)` not `isAtCommit(master)`.  This requires a
`git fetch` to be accurate.  I think it requires more work to do the
`isAtCommit` than to just always sync.

The right answer, I think, is to map the commit-ish to an actual sha1
at some point, and then use that sha1 in _setupWebapp.  I've left that
a TODO for now.

Auditors: amos, chris, colin

Test plan:
Fingers crossed