Write a version of timeout(1) that times out based on stdout.
Summary:
We have a bunch of jenkins jobs that had a 'die if the script has no
output in 20m'.  In the new jenkins pipeline world, that functionality
isn't exported, so I had to write it myself.

The basic implementation is based on subprocess.py's communicate()
logic, which uses select().  The main difference is I added a timeout
parameter.  I also had to change the logic because communicate()
assumes the input comes in via a string whereas I wanted to pass
through the input that came in on my stdin.  I *think* I got that
right in a way that doesn't cause a bunch of busy-polling.

Fix up support for `-k`.

I realized that `-k` should support duration units just like the
normal duration, so I changed the parsing to handle that.

I then managed to figure out a test case, and tested the
functionality.

Test Plan:
I ran the following:
   % ./timeout_output.py 5 sh -c 'echo hi; sleep 3; echo lo; sleep 3; echo foo; sleep 4; echo bar'
   hi
   lo
   foo
   bar
and it succeeded with rc 0.  I then ran:
   ./timeout_output.py 5 sh -c 'echo hi; sleep 3; echo lo; sleep 3; echo foo; sleep 6; echo bar'
   hi
   lo
   foo
and it failed with rc 124.

I then went to webapp and ran
   ../jenkins-tools/timeout_output.py 1m make quickcheck
and it succeeded.  But when I ran
   ../jenkins-tools/timeout_output.py 1s make quickcheck
it failed in the `npm install` step (which takes about 5s before it
produces any output).

   time ./timeout_output.py -k 2 2 python -c 'import signal, time; signal.signal(signal.SIGTERM, signal.SIG_IGN); time.sleep(100)'

This took 4 seconds to run and when I did a `ps` afterwards I did not
see any `python` processes.

On the other hand, when I ran:
   time ./timeout_output.py 2 python -c 'import signal, time; signal.signal(signal.SIGTERM, signal.SIG_IGN); time.sleep(100)'
it took 2 seconds to run and I did see `python` when I ran `ps`.

When I ran
   time ./timeout_output.py -k 2 2 python -c 'import time; time.sleep(100)'
it took 2 seconds to run and `ps` did not show any `python`, so the
easy case still works too.

Reviewers: amos, chris, colin

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D33496