Add back in a safer version of isAtCommit().
Summary:
This version uses a sentinel-file which is updated every time there's
a sync.  If the sentinel file indicates that *this jenkins build* has
already done a sync of the given repo to the given commit, we don't
try to do it again.

Extract out the build-tag filename to its own function.

I had a typo where the name was one thing in one function and another
thing in another function.  That's not going to work...  Having a
helper function makes that kind of bug impossible.

In wasSyncedTo, don't consider safeSyncTo the same as safeSyncToOrigin.

Suppose we do:
    safeSyncToOrigin(repo, branchname);
    safeSyncTo(repo, branchname);

Should that second call be a noop because we're already synced to the
given branch?  Not exactly: the first syncs the branch to
origin/branchname while the second syncs to branchname, which are
//probably// the same but at the least have different origins which
could behave differently.  It's safest to just have the second one not
be a noop.  It's unlikely people will do the above in practice.

Test Plan:
Fingers crossed

(none new)

(none new)

Reviewers: chris, colin, amos

Reviewed By: colin, amos

Differential Revision: https://phabricator.khanacademy.org/D33497