Refactor e2e-test to use an outer-level `notify()` on failure.
Summary:
I've yet to come up with a really good way of sending notifications to
slack when the script itself fails (rather than whatever it's trying
to do).  For instance, if a git-checkout times out, we want to be able
to notify via slack that the job has failed.

I was doing that before by wrapping lots of things in a try/finally,
but it was messy and ugly.  So I'm now trying a technique where I just
wrap the whole script in a new `notify()` step.  It waits until the
job is done and then talks to slack appropriately depending on how
things went down.

It's still ugly -- I have to use a global var to suppress
slack-sending when appropriate, for instance -- but turns out ok once
I refactor the code to not have to indent everything another level for
notify().

The bigger problem is groovy has a bug where exceptions raised in a
`finally` are not properly caught by an outer `catch`.  This means, in
our case, that errors in analyzeResults() will not be captured
properly by the `catch` inside `notify`.  I've filed a bug
(https://issues.apache.org/jira/browse/GROOVY-8085) and hopefully this
will get fixed, but until then this won't work as well as it should.

Rename all the test-files to be more consistent.

They're now called test-<foo>.  They used to be called e2e-test-<foo>
or e2e_test_<foo> or selenium_test_<foo> and now they are all just
called test-<foo>.

I got rid of the `e2e` prefix since it's not really needed; we're
already the e2e-test job so the e2e-ness is obvious from the directory
structure already.  This will let us share this code more easily
between e2e tests and webapp (python) tests.

Test Plan:
None, but I'm very interested in how it works!  It may be hard to
fully test until something goes wrong with the script framework, such
as a timeout error.

Fingers crossed

Reviewers: amos, chris, colin

Reviewed By: amos, colin

Differential Revision: https://phabricator.khanacademy.org/D33531